search: greatly improve search performance and UI responsiveness
Active searches are now immediately cancelled when the search input changes.
This can easily safe a couple hundred milliseconds for complex queries,
because we don't need to wait for the old query to finish before the new
query is processed.

Since this also avoids lots of unecessary UI updates the UI doesn't
block nearly as much as before when the search input changes quickly
(e.g. while pressing backspace).