results: improve performance when rendering icons
Previously we were stat'ing the filesystem synchronously to query the file information to detect the proper icon to use. Now we just guess the icon from the file name, which is faster and won't block the UI, at the cost of having less accurate results.

The long term goal should be to query the file system asynchronously to detect the icon name.